# docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    required:
    - code
    - message
    - status
    type: object
  LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  LoginResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      id:
        type: string
    required:
    - access_token
    - email
    - id
    type: object
  MeResponse:
    properties:
      email:
        type: string
      id:
        type: string
    required:
    - email
    - id
    type: object
  MessageResponse:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  SignupRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      status:
        type: string
    required:
    - email
    - password
    type: object
info:
  contact:
    email: andy@d.foundation
    name: Andy
    url: https://d.foundation
  description: This is api document for APP API project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: APP API DOCUMENT
  version: v0.0.1
paths:
  /portal/auth/login:
    post:
      consumes:
      - application/json
      description: Login to portal by email
      parameters:
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Login to portal
      tags:
      - Auth
  /portal/auth/signup:
    post:
      consumes:
      - application/json
      description: Signup
      parameters:
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Signup
      tags:
      - Auth
  /portal/me:
    get:
      consumes:
      - application/json
      description: Retrieve my information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve my information
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"